Pre-review tidies and fixes
