+++ app/services/editions/publishing/FeastPublicationTarget.scala
@@ -1,6 +1,7 @@
 package services.editions.publishing
 import com.amazonaws.services.sns.AmazonSNS
 import com.amazonaws.services.sns.model.{MessageAttributeValue, PublishRequest}
+import conf.ApplicationConfiguration
 import model.editions.PublishableIssue
 import play.api.Configuration
 import play.api.libs.json.{Json, Writes}
@@ -16,18 +17,13 @@ object FeastPublicationTarget {
   type MessageType = MessageType.Value
 }
 
-class FeastPublicationTarget(snsClient:AmazonSNS, config: Configuration) extends PublicationTargetWithTransform[FeastAppModel.FeastAppCuration] {
-  import PublishedIssueFormatters._
-
+class FeastPublicationTarget(snsClient:AmazonSNS, config: ApplicationConfiguration) extends PublicationTargetWithTransform[FeastAppModel.FeastAppCuration] {
   override val transform: PublicationTransform[FeastAppModel.FeastAppCuration] = FeastAppTransform()
-
-  //FIXME: put in a better config key for this
-  lazy private val publicationTopic = config.get[String]("feast_app.topic_arn")
 
   private def createPublishRequest(content:String, messageType:FeastPublicationTarget.MessageType):PublishRequest = {
     new PublishRequest()
       .withMessage(content)
-      .withTopicArn(publicationTopic)
+      .withTopicArn(config.aws.feastAppPublicationTopic)
       .withMessageAttributes(Map(
         "timestamp"->new MessageAttributeValue().withDataType("Number").withStringValue(Instant.now().toEpochMilli.toString),
         "type"->new MessageAttributeValue().withDataType("String").withStringValue(messageType.toString),
