+++ conf/application.conf
@@ -124,7 +124,10 @@ PROD {
 
 faciatool.show_test_containers=true
 
-feast_app.topic_arn = "arn:aws:sns:eu-west-1:163592447864:feast-facia-test"
+feast_app.publication_topic = "arn:aws:sns:eu-west-1:163592447864:facia-db-CODE-FeastPublicationTopic-rWHyUm0ckOji"
+PROD {
+    feast_app.publication_topic = "NOT-SET-UP-YET"
+}
 include "local.conf"
 
 include file("/etc/gu/facia-tool.application.secrets.conf")
