+++ conf/local.conf
@@ -4,4 +4,3 @@ db.default {
     password = faciatool
     port = 4724
 }
-
